Update Reset Button Read Me
I've become aware that I soldered my transistor incorrectly. When referring to picture 4, the transistor orientation left to right is Collector, Base, Emmiter. The green ground wire is suppose to be soldered to the far right Emmiter, and not the left. So still follow the instructions above, just note that my solder is incorrect when using picture 4 as a reference. More info inside of Read Me